Temporarily allowing the mobile nav bar to close
The mobile navigation menu has issues with closing and opening. This javascript is only present on the index.php page for now. If we want to use this "bandaid" on other pages, we'll need to include this code on those pages.

Long-term, we need an actual solution to the mobile menu not closing properly. Probably want to discuss with Dr. Jang.